+++ src/interfaces/http.rs
@@ -4,7 +4,7 @@
 //! ```toml
 //! [[server.interface]]
 //! type = "http"
-//! host = "127.0.0.1"
+//! host = "http://localhost"
 //! port = 8080
 //! ```
 //!
@@ -27,6 +27,7 @@ use crate::Error;
 use async_trait::async_trait;
 use serde::Deserialize;
 use std::net::SocketAddr;
+use std::path::PathBuf;
 use tokio::sync::{broadcast, mpsc, watch};
 use url::{ParseError, Url};
 
@@ -41,6 +42,8 @@ pub struct HttpSocketInterface {
     host: Url,
     tls: bool,
     port: u16,
+    tls_cert_path: Option<PathBuf>,
+    tls_key_path: Option<PathBuf>,
 }
 
 /// Data structure to represent the HTTP Socket [`InterfaceConfig`].
@@ -50,15 +53,19 @@ pub(crate) struct HttpSocketConfigFile {
     pub host: String,
     pub tls: bool,
     pub port: i64,
+    pub tls_cert_path: Option<String>,
+    pub tls_key_path: Option<String>,
 }
 
 impl HttpSocketInterface {
     /// Create a new `HttpSocketInterface`.
-    pub fn new(host_url: &Url) -> Self {
+    pub fn new<P: AsRef<str>>(host_url: &Url, cert_path: Option<P>, key_path: Option<P>) -> Self {
         let host = host_url.clone();
         let tls = host.scheme().eq_ignore_ascii_case(HTTPS);
         let port = host.port().unwrap_or(DEFAULT_PORT);
-        Self { host, tls, port }
+        let tls_cert_path = cert_path.map(|p| PathBuf::from(p.as_ref()));
+        let tls_key_path = key_path.map(|p| PathBuf::from(p.as_ref()));
+        Self { host, tls, port, tls_cert_path, tls_key_path }
     }
 
     /// Return the IP address.
@@ -75,17 +82,38 @@ impl HttpSocketInterface {
     pub fn port(&self) -> u16 {
         self.port
     }
+
+    /// Return if interface should use TLS
+    pub fn tls(&self) -> bool {
+        self.tls
+    }
+
+    /// Return path the TLS certificate
+    pub fn tls_cert_path(&self) -> &Option<PathBuf> {
+        &self.tls_cert_path
+    }
+
+    /// Return path the TLS private key
+    pub fn tls_key_path(&self) -> &Option<PathBuf> {
+        &self.tls_key_path
+    }
 }
 
 impl Default for HttpSocketConfigFile {
     fn default() -> Self {
-        Self { host: LOCALHOST.into(), tls: false, port: DEFAULT_PORT as i64 }
+        Self { host: LOCALHOST.into(), tls: false, port: DEFAULT_PORT as i64, tls_cert_path: None, tls_key_path: None }
     }
 }
 
 impl Default for HttpSocketInterface {
     fn default() -> Self {
-        Self { host: Url::parse(LOCALHOST).unwrap(), tls: false, port: DEFAULT_PORT }
+        Self {
+            host: Url::parse(LOCALHOST).unwrap(),
+            tls: false,
+            port: DEFAULT_PORT,
+            tls_cert_path: None,
+            tls_key_path: None,
+        }
     }
 }
 
@@ -102,8 +130,9 @@ impl TryFrom<&HttpSocketConfigFile> for HttpSocketInterface {
             false => url.set_scheme(HTTP),
         }
         .unwrap();
+
         url.set_port(Some(value.port as u16)).unwrap();
-        Ok(Self { host: url, tls: value.tls, port: value.port as u16 })
+        Ok(HttpSocketInterface::new(&url, value.tls_cert_path.as_ref(), value.tls_key_path.as_ref()))
     }
 }
 
@@ -120,13 +149,21 @@ impl Interface for HttpSocketInterface {
     async fn receive(&self, interface_tx: mpsc::Sender<String>, shutdown: watch::Receiver<bool>) -> Result<(), Error> {
         use crate::interfaces::http::http_server::start_monitoring;
 
-        for socket in self.sockets()? {
-            let itx = interface_tx.clone();
-            let srx = shutdown.clone();
-            tokio::spawn(async move { start_monitoring(itx, srx, socket).await });
+        if self.tls && (self.tls_cert_path().is_none() || self.tls_cert_path().is_none()) {
+            Err(Error::InvalidInterfaceConfiguration(
+                "Both tls_cert_path and tls_cert_path must be provided for a TLS server".into(),
+            ))
+        } else {
+            for socket in self.sockets()? {
+                let tls = self.tls;
+                let itx = interface_tx.clone();
+                let srx = shutdown.clone();
+                let cert_path = self.tls_cert_path.clone();
+                let key_path = self.tls_key_path.clone();
+                tokio::spawn(async move { start_monitoring(itx, srx, socket, tls, cert_path, key_path).await });
+            }
+            Ok(())
         }
-
-        Ok(())
     }
 
     #[cfg(not(feature = "http-server"))]